+++ gagu-front-end/src/types/index.ts
@@ -7,6 +7,7 @@ export * from './tunnel.type'
 export * from './setting.type'
 
 export type PageType = 'PENDING' | 'login' | 'desktop' | 'explore' | 'touch' | 'sharing'
+export type PublicImageName = 'bg-desktop' | 'bg-sharing' | 'favicon'
 
 export interface IResponse<T> {
   success: boolean